Add some stuff to EventBuffer

+++ ScoreFollower/ScoreFollower/types.h
@@ -6,7 +6,7 @@ namespace cf {
 namespace ScoreFollower {
 
 // score_time_t is in nanoseconds
-typedef unsigned long score_time_t;
+typedef long long score_time_t;
 
 typedef cf::timestamp_t real_time_t;
